Fix subtitle (date) on report card view
