+++ Tella/Domain/Entity/Nextcloud/NextcloudReport.swift
@@ -36,7 +36,9 @@ class NextcloudReport: BaseReport {
                    createdDate: createdDate,
                    updatedDate: updatedDate,
                    status: status,
-                   vaultFiles: vaultFiles)
+                   vaultFiles: vaultFiles,
+                   serverId: self.server?.id)
+
     }
     
     override func encode(to encoder: Encoder) throws {
